基本完成后端编写和测试，完成API文档编写

+++ apexflow_server/src/main/java/com/apex/core/service/AfterSalesService.java
@@ -102,7 +102,7 @@ public class AfterSalesService {
         result.setData(afterSalesList);
         result.setCurrentPage(page);
         result.setPageSize(pageSize);
-        result.setTotalCount(totalCount);
+        result.setTotalCount((int) totalCount);
 
         logger.info("After sales list retrieved. Total: {}, Current page: {}", totalCount, page);
         return result;
@@ -127,7 +127,7 @@ public class AfterSalesService {
         result.setData(afterSalesList);
         result.setCurrentPage(page);
         result.setPageSize(pageSize);
-        result.setTotalCount(totalCount);
+        result.setTotalCount((int) totalCount);
 
         logger.info("After sales list by status retrieved. Status: {}, Count: {}", status, afterSalesList.size());
         return result;
